#19701: Move conv2d tests from all post-commit pipeline to nightly pipeline
